third_party/blink/web_tests/platform/generic/external/wpt/service-workers/service-worker/dedicated-worker-service-worker-interception.https-expected.txt

This is a testharness.js-based test.
FAIL Top-level module loading should be intercepted by a service worker. assert_equals: expected "LOADED_FROM_SERVICE_WORKER" but got "LOADED_FROM_NETWORK"
FAIL Static import should be intercepted by a service worker. assert_equals: expected "LOADED_FROM_SERVICE_WORKER" but got "LOADED_FROM_NETWORK"
FAIL Dynamic import should be intercepted by a service worker. assert_equals: expected "LOADED_FROM_SERVICE_WORKER" but got "LOADED_FROM_NETWORK"
Harness: the test ran to completion.
